fix comments in schedules and move functions to utils.py
