+++ pipelines/datasets/br_anatel_banda_larga_fixa/schedules.py
@@ -13,10 +13,10 @@ from pipelines.constants import constants
 anatel_microdados = Schedule(
     clocks=[
         CronClock(
-            cron="0 15 * * *",  # ! goes execute every day 12 pm
-            start_date=datetime(2023, 4, 3, 7, 5, 0),  # ! first date that the pipeline executed
+            cron="0 15 * * *",
+            start_date=datetime(2023, 4, 3, 7, 5, 0),
             labels=[
-                constants.BASEDOSDADOS_PROD_AGENT_LABEL.value,  # ! label of identify (ex: basedosdados-dev)
+                constants.BASEDOSDADOS_PROD_AGENT_LABEL.value,
             ],
             parameter_defaults={
                 "dataset_id": "br_anatel_banda_larga_fixa",
@@ -33,10 +33,10 @@ anatel_microdados = Schedule(
 anatel_densidade_municipio = Schedule(
     clocks=[
         CronClock(
-            cron="0 16 * * *",  # ! goes execute every day 12 pm
-            start_date=datetime(2023, 4, 3, 7, 5, 0),  # ! first date that the pipeline executed
+            cron="0 16 * * *",
+            start_date=datetime(2023, 4, 3, 7, 5, 0),
             labels=[
-                constants.BASEDOSDADOS_PROD_AGENT_LABEL.value,  # ! label of identify (ex: basedosdados-dev)
+                constants.BASEDOSDADOS_PROD_AGENT_LABEL.value,
             ],
             parameter_defaults={
                 "dataset_id": "br_anatel_banda_larga_fixa",
@@ -53,10 +53,10 @@ anatel_densidade_municipio = Schedule(
 anatel_densidade_brasil = Schedule(
     clocks=[
         CronClock(
-            cron="0 17 * * *",  # ! goes execute every day 12 pm
-            start_date=datetime(2023, 4, 3, 7, 5, 0),  # ! first date that the pipeline executed
+            cron="0 17 * * *",
+            start_date=datetime(2023, 4, 3, 7, 5, 0),
             labels=[
-                constants.BASEDOSDADOS_PROD_AGENT_LABEL.value,  # ! label of identify (ex: basedosdados-dev)
+                constants.BASEDOSDADOS_PROD_AGENT_LABEL.value,
             ],
             parameter_defaults={
                 "dataset_id": "br_anatel_banda_larga_fixa",
@@ -73,10 +73,10 @@ anatel_densidade_brasil = Schedule(
 anatel_densidade_uf = Schedule(
     clocks=[
         CronClock(
-            cron="0 18 * * *",  # ! goes execute every day 12 pm
-            start_date=datetime(2023, 4, 3, 7, 5, 0),  # ! first date that the pipeline executed
+            cron="0 18 * * *",
+            start_date=datetime(2023, 4, 3, 7, 5, 0),
             labels=[
-                constants.BASEDOSDADOS_PROD_AGENT_LABEL.value,  # ! label of identify (ex: basedosdados-dev)
+                constants.BASEDOSDADOS_PROD_AGENT_LABEL.value,
             ],
             parameter_defaults={
                 "dataset_id": "br_anatel_banda_larga_fixa",
